systems: prevent internal error when mode/ips not available
After BMC reboot it can sometimes take a while for the power mode or
idle power status parameters to become available. This change will
prevent returning an error when there is a timeout trying to read them.
This will allow the other systems data to be returned.systems: prevent
internal error when mode/ips not available

After BMC reboot it can sometimes take a while for the power mode or
idle power status parameters to become available. This change will
prevent returning an error when there is a timeout trying to read them.
This will allow the other systems data to be returned.

Signed-off-by: Chris Cain <cjcain@us.ibm.com>
Change-Id: I65f7e2f4ccf946ce108ca10f72cb81dd0e8a7fd2